fix checkout qty to zero delete item

+++ src/shoppingCart/checkout.ts
@@ -605,22 +605,24 @@ Plugin.create(import.meta.url, (glitter, editMode) => {
                                                                             <div class="d-flex" style="">
                                                                                 <img style="width: 24px;height: 24px;" src="${new URL('../img/component/minusCircle.svg', import.meta.url)}" onclick="${gvc.event(() => {
                                                                     item.qty--;
-                                                                    
-                                                                    dialog.confirm("確定要刪除嘛?", (result) => {
-                                                                        if (result == true) {
-                                                                            item.deleteEvent()
-                                                                            category.item.splice(itemIndex, 1);
+                                                                    if (item.qty == 0){
+                                                                        dialog.confirm("確定要刪除嘛?", (result) => {
+                                                                            if (result == true) {
+                                                                                item.deleteEvent()
+                                                                                category.item.splice(itemIndex, 1);
 
-                                                                            if (category.item.length == 0) {
-                                                                                let indexToRemove = widget.data.cartItem.findIndex((item: any) => item.category_id == category.category_id);
-                                                                                widget.data.cartItem.splice(indexToRemove, 1);
-                                                                                refreshCart();
-                                                                            } else {
-                                                                                refreshCart();
+                                                                                if (category.item.length == 0) {
+                                                                                    let indexToRemove = widget.data.cartItem.findIndex((item: any) => item.category_id == category.category_id);
+                                                                                    widget.data.cartItem.splice(indexToRemove, 1);
+                                                                                    refreshCart();
+                                                                                } else {
+                                                                                    refreshCart();
+                                                                                }
+
                                                                             }
-
-                                                                        }
-                                                                    })
+                                                                        })
+                                                                    }
+                                                                    
                                                                     
                                                                 })}">
                                                                     ${gvc.bindView({
